+++ graft/coreth/plugin/evm/config/config.go
@@ -244,6 +244,9 @@ type Config struct {
 
 	// RPC settings
 	HttpBodyLimit uint64 `json:"http-body-limit"`
+
+	// Database Scheme
+	StateScheme string `json:"state-scheme"`
 }
 
 // TxPoolConfig contains the transaction pool config to be passed