+++ graft/coreth/plugin/evm/vm_test.go
@@ -47,6 +47,7 @@ import (
 	atomictxpool "github.com/ava-labs/coreth/plugin/evm/atomic/txpool"
 	atomicvm "github.com/ava-labs/coreth/plugin/evm/atomic/vm"
 	"github.com/ava-labs/coreth/plugin/evm/config"
+	"github.com/ava-labs/coreth/plugin/evm/customrawdb"
 	"github.com/ava-labs/coreth/plugin/evm/customtypes"
 	"github.com/ava-labs/coreth/plugin/evm/extension"
 	"github.com/ava-labs/coreth/plugin/evm/header"
@@ -58,6 +59,7 @@ import (
 	"github.com/ava-labs/coreth/utils"
 	"github.com/ava-labs/coreth/utils/utilstest"
 	"github.com/ava-labs/libevm/common"
+	"github.com/ava-labs/libevm/core/rawdb"
 	"github.com/ava-labs/libevm/core/types"
 	"github.com/ava-labs/libevm/log"
 	"github.com/ava-labs/libevm/rlp"
@@ -68,6 +70,7 @@ import (
 )
 
 var (
+	schemes          = []string{rawdb.HashScheme, customrawdb.FirewoodScheme}
 	initialBaseFee   = big.NewInt(ap3.InitialBaseFee)
 	testKeys         = secp256k1.TestKeys()[:3]
 	testEthAddrs     []common.Address // testEthAddrs[i] corresponds to testKeys[i]
@@ -253,6 +256,19 @@ func newVM(t *testing.T, config testVMConfig) *testVM {
 	}
 }
 
+// Firewood cannot yet be run with an empty config.
+func getConfig(scheme, otherConfig string) string {
+	innerConfig := otherConfig
+	if scheme == customrawdb.FirewoodScheme {
+		if len(innerConfig) > 0 {
+			innerConfig += ", "
+		}
+		innerConfig += fmt.Sprintf(`"state-scheme": "%s", "snapshot-cache": 0, "pruning-enabled": true, "state-sync-enabled": false`, customrawdb.FirewoodScheme)
+	}
+
+	return fmt.Sprintf(`{%s}`, innerConfig)
+}
+
 func (vm *testVM) WaitForEvent(ctx context.Context) commonEng.Message {
 	msg, err := vm.vm.WaitForEvent(ctx)
 	require.NoError(vm.t, err)
@@ -378,6 +394,14 @@ func TestVMContinuousProfiler(t *testing.T) {
 }
 
 func TestVMUpgrades(t *testing.T) {
+	for _, scheme := range schemes {
+		t.Run(scheme, func(t *testing.T) {
+			testVMUpgrades(t, scheme)
+		})
+	}
+}
+
+func testVMUpgrades(t *testing.T, scheme string) {
 	genesisTests := []struct {
 		fork             upgradetest.Fork
 		expectedGasPrice *big.Int
@@ -419,12 +443,14 @@ func TestVMUpgrades(t *testing.T) {
 			expectedGasPrice: big.NewInt(0),
 		},
 	}
+
 	for _, test := range genesisTests {
 		t.Run(test.fork.String(), func(t *testing.T) {
 			require := require.New(t)
 
 			vm := newVM(t, testVMConfig{
-				fork: &test.fork,
+				fork:       &test.fork,
+				configJSON: getConfig(scheme, ""),
 			}).vm
 			defer func() {
 				require.NoError(vm.Shutdown(context.Background()))
@@ -448,6 +474,14 @@ func TestVMUpgrades(t *testing.T) {
 }
 
 func TestImportMissingUTXOs(t *testing.T) {
+	for _, scheme := range schemes {
+		t.Run(scheme, func(t *testing.T) {
+			testImportMissingUTXOs(t, scheme)
+		})
+	}
+}
+
+func testImportMissingUTXOs(t *testing.T, scheme string) {
 	// make a VM with a shared memory that has an importable UTXO to build a block
 	importAmount := uint64(50000000)
 	fork := upgradetest.ApricotPhase2
@@ -456,6 +490,7 @@ func TestImportMissingUTXOs(t *testing.T) {
 		utxos: map[ids.ShortID]uint64{
 			testShortIDAddrs[0]: importAmount,
 		},
+		configJSON: getConfig(scheme, ""),
 	})
 	defer func() {
 		require.NoError(t, tvm1.vm.Shutdown(context.Background()))
@@ -470,7 +505,8 @@ func TestImportMissingUTXOs(t *testing.T) {
 
 	// make another VM which is missing the UTXO in shared memory
 	vm2 := newVM(t, testVMConfig{
-		fork: &fork,
+		fork:       &fork,
+		configJSON: getConfig(scheme, ""),
 	}).vm
 	defer func() {
 		require.NoError(t, vm2.Shutdown(context.Background()))
@@ -490,6 +526,14 @@ func TestImportMissingUTXOs(t *testing.T) {
 // Simple test to ensure we can issue an import transaction followed by an export transaction
 // and they will be indexed correctly when accepted.
 func TestIssueAtomicTxs(t *testing.T) {
+	for _, scheme := range schemes {
+		t.Run(scheme, func(t *testing.T) {
+			testIssueAtomicTxs(t, scheme)
+		})
+	}
+}
+
+func testIssueAtomicTxs(t *testing.T, scheme string) {
 	importAmount := uint64(50000000)
 	fork := upgradetest.ApricotPhase2
 	tvm := newVM(t, testVMConfig{
@@ -497,6 +541,7 @@ func TestIssueAtomicTxs(t *testing.T) {
 		utxos: map[ids.ShortID]uint64{
 			testShortIDAddrs[0]: importAmount,
 		},
+		configJSON: getConfig(scheme, ""),
 	})
 	defer func() {
 		if err := tvm.vm.Shutdown(context.Background()); err != nil {
@@ -600,11 +645,19 @@ func TestIssueAtomicTxs(t *testing.T) {
 }
 
 func TestBuildEthTxBlock(t *testing.T) {
+	for _, scheme := range schemes {
+		t.Run(scheme, func(t *testing.T) {
+			testBuildEthTxBlock(t, scheme)
+		})
+	}
+}
+
+func testBuildEthTxBlock(t *testing.T, scheme string) {
 	importAmount := uint64(20000000)
 	fork := upgradetest.ApricotPhase2
 	tvm := newVM(t, testVMConfig{
 		fork:       &fork,
-		configJSON: `{"pruning-enabled":true}`,
+		configJSON: getConfig(scheme, `"pruning-enabled":true`),
 		utxos: map[ids.ShortID]uint64{
 			testShortIDAddrs[0]: importAmount,
 		},
@@ -720,13 +773,14 @@ func TestBuildEthTxBlock(t *testing.T) {
 	restartedVM := atomicvm.WrapVM(&VM{})
 	newCTX := snowtest.Context(t, snowtest.CChainID)
 	newCTX.NetworkUpgrades = upgradetest.GetConfig(fork)
+	newCTX.ChainDataDir = tvm.vm.ctx.ChainDataDir
 	if err := restartedVM.Initialize(
 		context.Background(),
 		newCTX,
 		tvm.db,
 		[]byte(genesisJSON(forkToChainConfig[fork])),
 		[]byte(""),
-		[]byte(`{"pruning-enabled":true}`),
+		[]byte(getConfig(scheme, `"pruning-enabled":true`)),
 		[]*commonEng.Fx{},
 		nil,
 	); err != nil {
@@ -745,7 +799,7 @@ func TestBuildEthTxBlock(t *testing.T) {
 	}
 }
 
-func testConflictingImportTxs(t *testing.T, fork upgradetest.Fork) {
+func testConflictingImportTxs(t *testing.T, fork upgradetest.Fork, scheme string) {
 	importAmount := uint64(10000000)
 	tvm := newVM(t, testVMConfig{
 		fork: &fork,
@@ -754,6 +808,7 @@ func testConflictingImportTxs(t *testing.T, fork upgradetest.Fork) {
 			testShortIDAddrs[1]: importAmount,
 			testShortIDAddrs[2]: importAmount,
 		},
+		configJSON: getConfig(scheme, ""),
 	})
 	defer func() {
 		if err := tvm.vm.Shutdown(context.Background()); err != nil {
@@ -927,9 +982,16 @@ func testConflictingImportTxs(t *testing.T, fork upgradetest.Fork) {
 }
 
 func TestReissueAtomicTxHigherGasPrice(t *testing.T) {
-	kc := secp256k1fx.NewKeychain(testKeys...)
+	for _, scheme := range schemes {
+		t.Run(scheme, func(t *testing.T) {
+			testReissueAtomicTxHigherGasPrice(t, scheme)
+		})
+	}
+}
 
-	for name, issueTxs := range map[string]func(t *testing.T, vm *atomicvm.VM, sharedMemory *avalancheatomic.Memory) (issued []*atomic.Tx, discarded []*atomic.Tx){
+func testReissueAtomicTxHigherGasPrice(t *testing.T, scheme string) {
+	kc := secp256k1fx.NewKeychain(testKeys...)
+	tests := map[string]func(t *testing.T, vm *atomicvm.VM, sharedMemory *avalancheatomic.Memory) (issued []*atomic.Tx, discarded []*atomic.Tx){
 		"single UTXO override": func(t *testing.T, vm *atomicvm.VM, sharedMemory *avalancheatomic.Memory) (issued []*atomic.Tx, evicted []*atomic.Tx) {
 			utxo, err := addUTXO(sharedMemory, vm.Ctx, ids.GenerateTestID(), 0, vm.Ctx.AVAXAssetID, units.Avax, testShortIDAddrs[0])
 			if err != nil {
@@ -1030,11 +1092,13 @@ func TestReissueAtomicTxHigherGasPrice(t *testing.T) {
 
 			return []*atomic.Tx{reissuanceTx2}, []*atomic.Tx{importTx1, importTx2}
 		},
-	} {
+	}
+	for name, issueTxs := range tests {
 		t.Run(name, func(t *testing.T) {
 			fork := upgradetest.ApricotPhase5
 			tvm := newVM(t, testVMConfig{
-				fork: &fork,
+				fork:       &fork,
+				configJSON: getConfig(scheme, `"pruning-enabled":true`),
 			})
 			issuedTxs, evictedTxs := issueTxs(t, tvm.atomicVM, tvm.atomicMemory)
 
@@ -1060,7 +1124,11 @@ func TestConflictingImportTxsAcrossBlocks(t *testing.T) {
 		upgradetest.ApricotPhase5,
 	} {
 		t.Run(fork.String(), func(t *testing.T) {
-			testConflictingImportTxs(t, fork)
+			for _, scheme := range schemes {
+				t.Run(scheme, func(t *testing.T) {
+					testConflictingImportTxs(t, fork, scheme)
+				})
+			}
 		})
 	}
 }
@@ -1076,13 +1144,21 @@ func TestConflictingImportTxsAcrossBlocks(t *testing.T) {
 //	    |
 //	    D
 func TestSetPreferenceRace(t *testing.T) {
+	for _, scheme := range schemes {
+		t.Run(scheme, func(t *testing.T) {
+			testSetPreferenceRace(t, scheme)
+		})
+	}
+}
+
+func testSetPreferenceRace(t *testing.T, scheme string) {
 	// Create two VMs which will agree on block A and then
 	// build the two distinct preferred chains above
 	importAmount := uint64(1000000000)
 	fork := upgradetest.NoUpgrades
 	tvmConfig := testVMConfig{
 		fork:       &fork,
-		configJSON: `{"pruning-enabled":true}`,
+		configJSON: getConfig(scheme, `"pruning-enabled":true`),
 		utxos: map[ids.ShortID]uint64{
 			testShortIDAddrs[0]: importAmount,
 		},
@@ -1302,6 +1378,14 @@ func TestSetPreferenceRace(t *testing.T) {
 }
 
 func TestConflictingTransitiveAncestryWithGap(t *testing.T) {
+	for _, scheme := range schemes {
+		t.Run(scheme, func(t *testing.T) {
+			testConflictingTransitiveAncestryWithGap(t, scheme)
+		})
+	}
+}
+
+func testConflictingTransitiveAncestryWithGap(t *testing.T, scheme string) {
 	key := utilstest.NewKey(t)
 
 	key0 := testKeys[0]
@@ -1319,6 +1403,7 @@ func TestConflictingTransitiveAncestryWithGap(t *testing.T) {
 			addr0: importAmount,
 			addr1: importAmount,
 		},
+		configJSON: getConfig(scheme, ""),
 	})
 	defer func() {
 		if err := tvm.vm.Shutdown(context.Background()); err != nil {
@@ -1432,9 +1517,18 @@ func TestConflictingTransitiveAncestryWithGap(t *testing.T) {
 }
 
 func TestBonusBlocksTxs(t *testing.T) {
+	for _, scheme := range schemes {
+		t.Run(scheme, func(t *testing.T) {
+			testBonusBlocksTxs(t, scheme)
+		})
+	}
+}
+
+func testBonusBlocksTxs(t *testing.T, scheme string) {
 	fork := upgradetest.NoUpgrades
 	tvm := newVM(t, testVMConfig{
-		fork: &fork,
+		fork:       &fork,
+		configJSON: getConfig(scheme, ""),
 	})
 	defer func() {
 		if err := tvm.vm.Shutdown(context.Background()); err != nil {
@@ -1532,11 +1626,19 @@ func TestBonusBlocksTxs(t *testing.T) {
 // accept block C, which should be an orphaned block at this point and
 // get rejected.
 func TestReorgProtection(t *testing.T) {
+	for _, scheme := range schemes {
+		t.Run(scheme, func(t *testing.T) {
+			testReorgProtection(t, scheme)
+		})
+	}
+}
+
+func testReorgProtection(t *testing.T, scheme string) {
 	importAmount := uint64(1000000000)
 	fork := upgradetest.NoUpgrades
 	tvmConfig := testVMConfig{
 		fork:       &fork,
-		configJSON: `{"pruning-enabled":false}`,
+		configJSON: getConfig(scheme, `"pruning-enabled":false`),
 		utxos: map[ids.ShortID]uint64{
 			testShortIDAddrs[0]: importAmount,
 		},
@@ -1703,6 +1805,13 @@ func TestReorgProtection(t *testing.T) {
 //	 / \
 //	B   C
 func TestNonCanonicalAccept(t *testing.T) {
+	for _, scheme := range schemes {
+		t.Run(scheme, func(t *testing.T) {
+			testNonCanonicalAccept(t, scheme)
+		})
+	}
+}
+func testNonCanonicalAccept(t *testing.T, scheme string) {
 	importAmount := uint64(1000000000)
 	fork := upgradetest.NoUpgrades
 	tvmConfig := testVMConfig{
@@ -1710,6 +1819,7 @@ func TestNonCanonicalAccept(t *testing.T) {
 		utxos: map[ids.ShortID]uint64{
 			testShortIDAddrs[0]: importAmount,
 		},
+		configJSON: getConfig(scheme, ""),
 	}
 	tvm1 := newVM(t, tvmConfig)
 	tvm2 := newVM(t, tvmConfig)
@@ -1900,6 +2010,14 @@ func TestNonCanonicalAccept(t *testing.T) {
 //	    |
 //	    D
 func TestStickyPreference(t *testing.T) {
+	for _, scheme := range schemes {
+		t.Run(scheme, func(t *testing.T) {
+			testStickyPreference(t, scheme)
+		})
+	}
+}
+
+func testStickyPreference(t *testing.T, scheme string) {
 	importAmount := uint64(1000000000)
 	fork := upgradetest.NoUpgrades
 	tvmConfig := testVMConfig{
@@ -1907,6 +2025,7 @@ func TestStickyPreference(t *testing.T) {
 		utxos: map[ids.ShortID]uint64{
 			testShortIDAddrs[0]: importAmount,
 		},
+		configJSON: getConfig(scheme, ""),
 	}
 	tvm1 := newVM(t, tvmConfig)
 	tvm2 := newVM(t, tvmConfig)
@@ -2161,6 +2280,14 @@ func TestStickyPreference(t *testing.T) {
 //	    |
 //	    D
 func TestUncleBlock(t *testing.T) {
+	for _, scheme := range schemes {
+		t.Run(scheme, func(t *testing.T) {
+			testUncleBlock(t, scheme)
+		})
+	}
+}
+
+func testUncleBlock(t *testing.T, scheme string) {
 	importAmount := uint64(1000000000)
 	fork := upgradetest.NoUpgrades
 	tvmConfig := testVMConfig{
@@ -2168,6 +2295,7 @@ func TestUncleBlock(t *testing.T) {
 		utxos: map[ids.ShortID]uint64{
 			testShortIDAddrs[0]: importAmount,
 		},
+		configJSON: getConfig(scheme, ""),
 	}
 	tvm1 := newVM(t, tvmConfig)
 	tvm2 := newVM(t, tvmConfig)
@@ -2345,6 +2473,14 @@ func TestUncleBlock(t *testing.T) {
 // Regression test to ensure that a VM that is not able to parse a block that
 // contains no transactions.
 func TestEmptyBlock(t *testing.T) {
+	for _, scheme := range schemes {
+		t.Run(scheme, func(t *testing.T) {
+			testEmptyBlock(t, scheme)
+		})
+	}
+}
+
+func testEmptyBlock(t *testing.T, scheme string) {
 	importAmount := uint64(1000000000)
 	fork := upgradetest.NoUpgrades
 	tvm := newVM(t, testVMConfig{
@@ -2352,6 +2488,7 @@ func TestEmptyBlock(t *testing.T) {
 		utxos: map[ids.ShortID]uint64{
 			testShortIDAddrs[0]: importAmount,
 		},
+		configJSON: getConfig(scheme, ""),
 	})
 	defer func() {
 		if err := tvm.vm.Shutdown(context.Background()); err != nil {
@@ -2414,6 +2551,14 @@ func TestEmptyBlock(t *testing.T) {
 //	    |
 //	    D
 func TestAcceptReorg(t *testing.T) {
+	for _, scheme := range schemes {
+		t.Run(scheme, func(t *testing.T) {
+			testAcceptReorg(t, scheme)
+		})
+	}
+}
+
+func testAcceptReorg(t *testing.T, scheme string) {
 	importAmount := uint64(1000000000)
 	fork := upgradetest.NoUpgrades
 	tvmConfig := testVMConfig{
@@ -2421,6 +2566,7 @@ func TestAcceptReorg(t *testing.T) {
 		utxos: map[ids.ShortID]uint64{
 			testShortIDAddrs[0]: importAmount,
 		},
+		configJSON: getConfig(scheme, ""),
 	}
 	tvm1 := newVM(t, tvmConfig)
 	tvm2 := newVM(t, tvmConfig)
@@ -2615,6 +2761,14 @@ func TestAcceptReorg(t *testing.T) {
 }
 
 func TestFutureBlock(t *testing.T) {
+	for _, scheme := range schemes {
+		t.Run(scheme, func(t *testing.T) {
+			testFutureBlock(t, scheme)
+		})
+	}
+}
+
+func testFutureBlock(t *testing.T, scheme string) {
 	importAmount := uint64(1000000000)
 	fork := upgradetest.NoUpgrades
 	tvm := newVM(t, testVMConfig{
@@ -2622,6 +2776,7 @@ func TestFutureBlock(t *testing.T) {
 		utxos: map[ids.ShortID]uint64{
 			testShortIDAddrs[0]: importAmount,
 		},
+		configJSON: getConfig(scheme, ""),
 	})
 	defer func() {
 		if err := tvm.vm.Shutdown(context.Background()); err != nil {
@@ -2678,6 +2833,14 @@ func TestFutureBlock(t *testing.T) {
 // Regression test to ensure we can build blocks if we are starting with the
 // Apricot Phase 1 ruleset in genesis.
 func TestBuildApricotPhase1Block(t *testing.T) {
+	for _, scheme := range schemes {
+		t.Run(scheme, func(t *testing.T) {
+			testBuildApricotPhase1Block(t, scheme)
+		})
+	}
+}
+
+func testBuildApricotPhase1Block(t *testing.T, scheme string) {
 	importAmount := uint64(1000000000)
 	fork := upgradetest.ApricotPhase1
 	tvm := newVM(t, testVMConfig{
@@ -2685,6 +2848,7 @@ func TestBuildApricotPhase1Block(t *testing.T) {
 		utxos: map[ids.ShortID]uint64{
 			testShortIDAddrs[0]: importAmount,
 		},
+		configJSON: getConfig(scheme, ""),
 	})
 	defer func() {
 		if err := tvm.vm.Shutdown(context.Background()); err != nil {
@@ -2791,6 +2955,14 @@ func TestBuildApricotPhase1Block(t *testing.T) {
 }
 
 func TestLastAcceptedBlockNumberAllow(t *testing.T) {
+	for _, scheme := range schemes {
+		t.Run(scheme, func(t *testing.T) {
+			testLastAcceptedBlockNumberAllow(t, scheme)
+		})
+	}
+}
+
+func testLastAcceptedBlockNumberAllow(t *testing.T, scheme string) {
 	importAmount := uint64(1000000000)
 	fork := upgradetest.NoUpgrades
 	tvm := newVM(t, testVMConfig{
@@ -2798,6 +2970,7 @@ func TestLastAcceptedBlockNumberAllow(t *testing.T) {
 		utxos: map[ids.ShortID]uint64{
 			testShortIDAddrs[0]: importAmount,
 		},
+		configJSON: getConfig(scheme, ""),
 	})
 	defer func() {
 		if err := tvm.vm.Shutdown(context.Background()); err != nil {
@@ -2863,6 +3036,14 @@ func TestLastAcceptedBlockNumberAllow(t *testing.T) {
 // that does not conflict. Accepts [blkB] and rejects [blkA], then asserts that the virtuous atomic
 // transaction in [blkA] is correctly re-issued into the atomic transaction mempool.
 func TestReissueAtomicTx(t *testing.T) {
+	for _, scheme := range schemes {
+		t.Run(scheme, func(t *testing.T) {
+			testReissueAtomicTx(t, scheme)
+		})
+	}
+}
+
+func testReissueAtomicTx(t *testing.T, scheme string) {
 	fork := upgradetest.ApricotPhase1
 	tvm := newVM(t, testVMConfig{
 		fork: &fork,
@@ -2870,6 +3051,7 @@ func TestReissueAtomicTx(t *testing.T) {
 			testShortIDAddrs[0]: 10000000,
 			testShortIDAddrs[1]: 10000000,
 		},
+		configJSON: getConfig(scheme, ""),
 	})
 	defer func() {
 		if err := tvm.vm.Shutdown(context.Background()); err != nil {
@@ -2960,9 +3142,18 @@ func TestReissueAtomicTx(t *testing.T) {
 }
 
 func TestAtomicTxFailsEVMStateTransferBuildBlock(t *testing.T) {
+	for _, scheme := range schemes {
+		t.Run(scheme, func(t *testing.T) {
+			testAtomicTxFailsEVMStateTransferBuildBlock(t, scheme)
+		})
+	}
+}
+
+func testAtomicTxFailsEVMStateTransferBuildBlock(t *testing.T, scheme string) {
 	fork := upgradetest.ApricotPhase1
 	tvm := newVM(t, testVMConfig{
-		fork: &fork,
+		fork:       &fork,
+		configJSON: getConfig(scheme, ""),
 	})
 	defer func() {
 		if err := tvm.vm.Shutdown(context.Background()); err != nil {
@@ -3010,9 +3201,18 @@ func TestAtomicTxFailsEVMStateTransferBuildBlock(t *testing.T) {
 }
 
 func TestBuildInvalidBlockHead(t *testing.T) {
+	for _, scheme := range schemes {
+		t.Run(scheme, func(t *testing.T) {
+			testBuildInvalidBlockHead(t, scheme)
+		})
+	}
+}
+
+func testBuildInvalidBlockHead(t *testing.T, scheme string) {
 	fork := upgradetest.ApricotPhase1
 	tvm := newVM(t, testVMConfig{
-		fork: &fork,
+		fork:       &fork,
+		configJSON: getConfig(scheme, ""),
 	})
 	defer func() {
 		if err := tvm.vm.Shutdown(context.Background()); err != nil {
@@ -3078,9 +3278,18 @@ func TestBuildInvalidBlockHead(t *testing.T) {
 // Regression test to ensure we can build blocks if we are starting with the
 // Apricot Phase 4 ruleset in genesis.
 func TestBuildApricotPhase4Block(t *testing.T) {
+	for _, scheme := range schemes {
+		t.Run(scheme, func(t *testing.T) {
+			testBuildApricotPhase4Block(t, scheme)
+		})
+	}
+}
+
+func testBuildApricotPhase4Block(t *testing.T, scheme string) {
 	fork := upgradetest.ApricotPhase4
 	tvm := newVM(t, testVMConfig{
-		fork: &fork,
+		fork:       &fork,
+		configJSON: getConfig(scheme, ""),
 	})
 	defer func() {
 		if err := tvm.vm.Shutdown(context.Background()); err != nil {
@@ -3250,9 +3459,18 @@ func TestBuildApricotPhase4Block(t *testing.T) {
 // Regression test to ensure we can build blocks if we are starting with the
 // Apricot Phase 5 ruleset in genesis.
 func TestBuildApricotPhase5Block(t *testing.T) {
+	for _, scheme := range schemes {
+		t.Run(scheme, func(t *testing.T) {
+			testBuildApricotPhase5Block(t, scheme)
+		})
+	}
+}
+
+func testBuildApricotPhase5Block(t *testing.T, scheme string) {
 	fork := upgradetest.ApricotPhase5
 	tvm := newVM(t, testVMConfig{
-		fork: &fork,
+		fork:       &fork,
+		configJSON: getConfig(scheme, ""),
 	})
 	defer func() {
 		if err := tvm.vm.Shutdown(context.Background()); err != nil {