+++ graft/coreth/RELEASES.md
@@ -8,6 +8,7 @@
 - Moves atomic operations from plugin/evm to plugin/evm/atomic and wraps the plugin/evm/VM in `atomicvm` to separate the atomic operations from the EVM execution.
 - Demoted unnecessary error log in `core/txpool/legacypool.go` to warning, displaying unexpected but valid behavior.
 - Removed the `snowman-api-enabled` flag and the corresponding API implementation.
+- Enable expermiental `state-scheme` flag to specify Firewood as a state database.
 
 ## [v0.15.1](https://github.com/ava-labs/coreth/releases/tag/v0.15.1)
 
@@ -37,6 +38,7 @@
 - Refactored trie_prefetcher.go to be structurally similar to [upstream](https://github.com/ethereum/go-ethereum/tree/v1.13.14).
 
 ## [v0.14.0](https://github.com/ava-labs/coreth/releases/tag/v0.14.0)
+
 - Minor version update to correspond to avalanchego v1.12.0 / Etna.
 - Remove unused historical opcodes CALLEX, BALANCEMC
 - Remove unused pre-AP2 handling of genesis contract
@@ -45,6 +47,7 @@
 - Update go version to 1.22
 
 ## [v0.13.8](https://github.com/ava-labs/coreth/releases/tag/v0.13.8)
+
 - Update geth dependency to v1.13.14
 - eupgrade: lowering the base fee to 1 nAVAX
 - eupgrade/cancun: verify no blobs in header
@@ -53,6 +56,7 @@
 - Remove cross-chain handlers
 
 ## [v0.13.7](https://github.com/ava-labs/coreth/releases/tag/v0.13.7)
+
 - Add EUpgrade base definitions
 - Remove Block Status
 - Fix and improve "GetBlockIDAtHeight"
@@ -60,11 +64,13 @@
 - Bump AvalancheGo to v1.11.10-prerelease
 
 ## [v0.13.6](https://github.com/ava-labs/coreth/releases/tag/v0.13.6)
+
 - rpc: truncate call error data logs
 - logging: remove path prefix (up to coreth@version/) from logged file names.
 - cleanup: removes pre-Durango scripts
 
 ## [v0.13.5](https://github.com/ava-labs/coreth/releases/tag/v0.13.5)
+
 - Bump AvalancheGo to v1.11.7
 - Bump golang version requirement to 1.21.12
 - Switches timestamp log back to "timestamp" (as was before v0.13.4)
@@ -73,6 +79,7 @@
 - Fix state sync crash bug
 
 ## [v0.13.4](https://github.com/ava-labs/coreth/releases/tag/v0.13.4)
+
 - Fixes snapshot use when state sync was explicitly enabled
 - Fixes v0.13.3 locking regression in async snapshot generation
 - Update go-ethereum to v1.13.8
@@ -81,6 +88,7 @@
 - "timestamp" in logs is changed to "t"
 
 ## [v0.13.3](https://github.com/ava-labs/coreth/releases/tag/v0.13.3)
+
 - Update go-ethereum to v1.13.2
 - Bump AvalancheGo to v1.11.5
 - Bump golang version requirement to 1.21.9
@@ -89,6 +97,7 @@
 - Testing improvements
 
 ## [v0.13.2](https://github.com/ava-labs/coreth/releases/tag/v0.13.2)
+
 - Integrate stake weighted gossip selection
 - Update go-ethereum to v1.12.2
 - Force precompile modules registration in ethclient